requirements.txt
Django==6.0.4
djangorestframework==3.17.1
django-filter==25.2
channels==4.3.2
daphne==4.2.1
drf-spectacular==0.29.0
django-cors-headers==4.9.0
